Clean up the `Login` component
