+++ app/renderer/views/CreatePortfolio/CreatePortfolio.scss
@@ -1,4 +1,4 @@
-@import '../styles/variables';
+@import '../../styles/variables';
 
 .CreatePortfolio {
 	.generated-seed-phrase-container {
