+++ app/renderer/views/CreatePortfolio/Step1.js
@@ -1,8 +1,8 @@
 import React from 'react';
-import Button from '../components/Button';
-import Input from '../components/Input';
-import LoginBackButton from '../components/LoginBackButton';
-import {createPortfolioContainer as container} from '../containers/CreatePortfolio';
+import Button from '../../components/Button';
+import Input from '../../components/Input';
+import LoginBackButton from '../../components/LoginBackButton';
+import {createPortfolioContainer as container} from '../../containers/CreatePortfolio';
 import './CreatePortfolio.scss';
 
 const CreatePortfolioStep1 = () => {
