+++ app/renderer/views/CreatePortfolio/Step3.js
@@ -1,8 +1,8 @@
 import React from 'react';
-import Button from '../components/Button';
-import TextArea from '../components/TextArea';
-import LoginBackButton from '../components/LoginBackButton';
-import {createPortfolioContainer as container} from '../containers/CreatePortfolio';
+import Button from '../../components/Button';
+import TextArea from '../../components/TextArea';
+import LoginBackButton from '../../components/LoginBackButton';
+import {createPortfolioContainer as container} from '../../containers/CreatePortfolio';
 import './CreatePortfolio.scss';
 
 const CreatePortfolioStep3 = () => {
