+++ app/renderer/views/CreatePortfolio/Step4.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import Success from '../components/Success';
+import Success from '../../components/Success';
 
 const CreatePortfolioStep4 = () => <Success>New portfolio added!</Success>;
