+++ app/renderer/views/ForgotPassword/ForgotPassword.scss
