+++ app/renderer/views/ForgotPassword/Step1.js
@@ -1,8 +1,8 @@
 import React from 'react';
-import Button from '../components/Button';
-import TextArea from '../components/TextArea';
-import LoginBackButton from '../components/LoginBackButton';
-import {forgotPasswordContainer as container} from '../containers/ForgotPassword';
+import Button from '../../components/Button';
+import TextArea from '../../components/TextArea';
+import LoginBackButton from '../../components/LoginBackButton';
+import {forgotPasswordContainer as container} from '../../containers/ForgotPassword';
 import './ForgotPassword.scss';
 
 const ForgotPasswordStep1 = () => {
