+++ app/renderer/views/ForgotPassword/Step2.js
@@ -1,10 +1,10 @@
 import React from 'react';
-import Button from '../components/Button';
-import Input from '../components/Input';
-import LoginBackButton from '../components/LoginBackButton';
-import {loginContainer} from '../containers/Login';
-import {forgotPasswordContainer as container} from '../containers/ForgotPassword';
-import avatar from '../avatar';
+import Button from '../../components/Button';
+import Input from '../../components/Input';
+import LoginBackButton from '../../components/LoginBackButton';
+import {loginContainer} from '../../containers/Login';
+import {forgotPasswordContainer as container} from '../../containers/ForgotPassword';
+import avatar from '../../avatar';
 import './ForgotPassword.scss';
 
 const ForgotPasswordStep2 = () => {
