+++ app/renderer/views/RestorePortfolio/Step3.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import Success from '../components/Success';
+import Success from '../../components/Success';
 
 const RestorePortfolioStep3 = () => <Success>New portfolio added!</Success>;
